+++ lib/core_ext/hash_with_indifferent_access.rb
@@ -1,9 +1,0 @@
-class Thor
-  module CoreExt #:nodoc:
-    class HashWithIndifferentAccess < ::Hash #:nodoc:
-      def has_key?(key)
-        super(convert_key(key))
-      end
-    end
-  end
-end