+++ lib/guard.rb
@@ -1,6 +1,3 @@
-require 'core_ext/hash_with_indifferent_access'
-#require 'pry'
-
 module Guard
 
   autoload :UI,           'guard/ui'
@@ -17,7 +14,7 @@ module Guard
     # initialize this singleton
     def setup(options = {})
       @options  = options
-      @listener = Listener.select_and_init((options[:watchdir]) ? File.expand_path(options[:watchdir]) : Dir.pwd)
+      @listener = Listener.select_and_init(@options[:watchdir] ? File.expand_path(@options[:watchdir]) : Dir.pwd)
       @guards   = []
 
       @options[:notify] && ENV["GUARD_NOTIFY"] != 'false' ? Notifier.turn_on : Notifier.turn_off